api: split globals and instantiation modules
Why:
- wago.go had grown to mix compile, const-expression, global metadata, and instantiation responsibilities.

What:
- Move global public types, metadata helpers, slot helpers, and exported-global accessors into globals.go.
- Move const-expression initializer helpers into constexpr.go.
- Move Instance lifecycle and instantiation code into instantiate.go.
- Keep compile, serialization, load, and run helpers in wago.go.

Proof:
- go test ./...

Docs:
- unchanged, internal mechanical split with no workflow impact.
