+++ wago.go
@@ -6,150 +6,13 @@ import (
 	"encoding/binary"
 	"encoding/gob"
 	"fmt"
-	"math"
 	"time"
-	"unsafe"
 
 	"github.com/wago-org/wago/src/core/compiler/backend/amd64"
 	"github.com/wago-org/wago/src/core/compiler/wasm"
-	"github.com/wago-org/wago/src/core/runtime"
 )
 
-// Value is a typed wasm call argument or result.
-type Value struct {
-	Type wasm.ValType
-	Bits uint64
-}
-
-func I32(v int32) Value   { return Value{wasm.I32, uint64(uint32(v))} }
-func I64(v int64) Value   { return Value{wasm.I64, uint64(v)} }
-func F32(v float32) Value { return Value{wasm.F32, uint64(math.Float32bits(v))} }
-func F64(v float64) Value { return Value{wasm.F64, math.Float64bits(v)} }
-
-func (v Value) AsI32() int32   { return int32(uint32(v.Bits)) }
-func (v Value) AsI64() int64   { return int64(v.Bits) }
-func (v Value) AsF32() float32 { return math.Float32frombits(uint32(v.Bits)) }
-func (v Value) AsF64() float64 { return math.Float64frombits(v.Bits) }
-
-func (v Value) String() string {
-	switch v.Type {
-	case wasm.I64:
-		return fmt.Sprintf("%d", v.AsI64())
-	case wasm.F32:
-		return fmt.Sprintf("%g", v.AsF32())
-	case wasm.F64:
-		return fmt.Sprintf("%g", v.AsF64())
-	default:
-		return fmt.Sprintf("%d", v.AsI32())
-	}
-}
-
-// HostFunc handles a void host import with one i32 argument.
-type HostFunc func(arg int32)
-
-// Imports supplies host functions and globals by "module.name" import key.
-type Imports struct {
-	Funcs   map[string]HostFunc
-	Globals map[string]GlobalImport
-}
-
-// GlobalImport is the initial value and type contract for an imported global.
-// Bits uses wasm's raw numeric encoding: i32/f32 consume the low 32 bits
-// (integer bits or IEEE-754 f32 bits), while i64/f64 consume all 64 bits
-// (integer bits or IEEE-754 f64 bits). InstantiateWithImports copies Bits into
-// an instance-local slot; the import is not retained or aliased for later
-// host-side mutation.
-type GlobalImport struct {
-	Type    wasm.ValType
-	Mutable bool
-	Bits    uint64
-}
-
-type FuncSig struct{ Params, Results []wasm.ValType }
-
-// ElemInit is active element-segment metadata. Base is the literal i32 offset.
-// When HasOffsetGlobal is true, OffsetGlobal names an imported immutable global
-// whose current instance slot is read during instantiation instead, after import
-// values have been copied into globals storage.
-type ElemInit struct {
-	Base            uint32
-	HasOffsetGlobal bool
-	OffsetGlobal    int
-	Funcs           []uint32
-}
-
-// DataInit is active data-segment metadata. Offset is the literal i32 offset.
-// When HasOffsetGlobal is true, OffsetGlobal names an imported immutable global
-// whose current instance slot is read during instantiation instead, after import
-// values have been copied into globals storage.
-type DataInit struct {
-	Offset          uint32
-	HasOffsetGlobal bool
-	OffsetGlobal    int
-	Bytes           []byte
-}
-
-// GlobalDef is the compact instantiate-time metadata for one wasm global.
-// Each instance stores one 8-byte slot per global; i32/f32 use the low 32 bits.
-// Bits is the literal initializer. When HasInitGlobal is true, InitGlobal names
-// an earlier imported immutable global whose value is copied into this global's
-// instance-local slot during instantiation; it is not a slot alias.
-type GlobalDef struct {
-	Type          wasm.ValType
-	Mutable       bool
-	Bits          uint64
-	HasInitGlobal bool
-	InitGlobal    int
-}
-
-// GlobalImportDef identifies one imported global slot in wasm global-index order.
-type GlobalImportDef struct {
-	Module  string
-	Name    string
-	Type    wasm.ValType
-	Mutable bool
-}
-
-// Compiled is emitted machine code plus instantiate-time metadata.
-type Compiled struct {
-	Code       []byte
-	Entry      []int          // entry offset per local function
-	Funcs      []FuncSig      // signature per local function
-	Imports    []string       // "module.name" per imported function
-	Exports    map[string]int // exported function name -> global function index
-	NumImports int
-
-	GlobalImports []GlobalImportDef // imported global slots, preceding local globals
-	Globals       []GlobalDef       // global slots in wasm global-index order
-	GlobalExports map[string]int    // exported global name -> global index
-
-	TableSize  int        // initial table length
-	FuncTypeID []uint32   // canonical signature id per global function index
-	Elems      []ElemInit // active element segments
-
-	Data []DataInit // active data segments (copied into linear memory at instantiate)
-}
-
 type Timings struct{ Decode, Validate, Compile time.Duration }
-
-// ImportedGlobalCount returns the number of imported globals at the front of
-// the wasm global-index space.
-func (c *Compiled) ImportedGlobalCount() int { return len(c.GlobalImports) }
-
-// LocalGlobalCount returns the number of module-defined globals.
-func (c *Compiled) LocalGlobalCount() int { return len(c.Globals) - len(c.GlobalImports) }
-
-// GlobalSlot maps a wasm global index to its byte offset in instance storage.
-func (c *Compiled) GlobalSlot(idx int) int { return idx * 8 }
-
-// ExportedGlobal returns metadata for a named exported global.
-func (c *Compiled) ExportedGlobal(name string) (GlobalDef, bool) {
-	idx, ok := c.GlobalExports[name]
-	if !ok || idx < 0 || idx >= len(c.Globals) {
-		return GlobalDef{}, false
-	}
-	return c.Globals[idx], true
-}
 
 // Compile decodes, validates, and compiles a wasm module to native code.
 func Compile(wasmBytes []byte) (*Compiled, error) {
@@ -274,131 +137,6 @@ func rejectUnsupportedGlobalTypes(m *wasm.Module) error {
 	return nil
 }
 
-func evalI32ConstExpr(b []byte) (uint32, error) {
-	v, err := evalConstExpr(b, wasm.I32)
-	if err != nil {
-		return 0, err
-	}
-	return uint32(v.Bits), nil
-}
-
-// constExprInit is the internal reusable form for MVP const expressions whose
-// literal bits are known at compile time unless an imported immutable global must
-// be read later, after import values are supplied during instantiation.
-type constExprInit struct {
-	Bits        uint64
-	GlobalIndex int
-}
-
-func (i constExprInit) GlobalRef() (int, bool) { return i.GlobalIndex, i.GlobalIndex >= 0 }
-
-type constExprResult struct {
-	Value
-	GlobalIndex int
-}
-
-func (r constExprResult) Init() constExprInit {
-	return constExprInit{Bits: r.Bits, GlobalIndex: r.GlobalIndex}
-}
-
-func applyGlobalInit(g *GlobalDef, init constExprInit) {
-	g.Bits = init.Bits
-	if idx, ok := init.GlobalRef(); ok {
-		g.HasInitGlobal = true
-		g.InitGlobal = idx
-	}
-}
-
-func applyElemOffset(e *ElemInit, init constExprInit) {
-	e.Base = uint32(init.Bits)
-	if idx, ok := init.GlobalRef(); ok {
-		e.HasOffsetGlobal = true
-		e.OffsetGlobal = idx
-	}
-}
-
-func applyDataOffset(d *DataInit, init constExprInit) {
-	d.Offset = uint32(init.Bits)
-	if idx, ok := init.GlobalRef(); ok {
-		d.HasOffsetGlobal = true
-		d.OffsetGlobal = idx
-	}
-}
-
-func evalConstExpr(b []byte, want wasm.ValType) (Value, error) {
-	res, err := evalConstExprWithModule(b, want, nil)
-	return res.Value, err
-}
-
-// evalConstExprWithModule intentionally stays narrower than full wasm validation:
-// wasm.Validate checks const-expression shape/type rules before compile reaches
-// here, while this helper decodes the supported MVP operators into
-// instantiate-time bits or clear unsupported-expression errors.
-func evalConstExprWithModule(b []byte, want wasm.ValType, m *wasm.Module) (constExprResult, error) {
-	r := wasm.NewReader(b)
-	op, err := r.Byte()
-	if err != nil {
-		return constExprResult{}, err
-	}
-	got := constExprResult{GlobalIndex: -1}
-	switch op {
-	case 0x41: // i32.const
-		v, err := r.I32()
-		if err != nil {
-			return constExprResult{}, err
-		}
-		got.Value = Value{Type: wasm.I32, Bits: uint64(uint32(v))}
-	case 0x42: // i64.const
-		v, err := r.I64()
-		if err != nil {
-			return constExprResult{}, err
-		}
-		got.Value = Value{Type: wasm.I64, Bits: uint64(v)}
-	case 0x43: // f32.const
-		bb, err := r.Bytes(4)
-		if err != nil {
-			return constExprResult{}, err
-		}
-		got.Value = Value{Type: wasm.F32, Bits: uint64(binary.LittleEndian.Uint32(bb))}
-	case 0x44: // f64.const
-		bb, err := r.Bytes(8)
-		if err != nil {
-			return constExprResult{}, err
-		}
-		got.Value = Value{Type: wasm.F64, Bits: binary.LittleEndian.Uint64(bb)}
-	case 0x23: // global.get
-		if m == nil {
-			return constExprResult{}, fmt.Errorf("unsupported const expression opcode 0x%02x", op)
-		}
-		x, err := r.U32()
-		if err != nil {
-			return constExprResult{}, err
-		}
-		gt, ok := m.GlobalType(x)
-		if !ok || int(x) >= m.ImportedGlobalCount() || gt.Mutable {
-			return constExprResult{}, fmt.Errorf("unsupported const expression global.get %d", x)
-		}
-		got.Value = Value{Type: gt.Val}
-		got.GlobalIndex = int(x)
-	default:
-		return constExprResult{}, fmt.Errorf("unsupported const expression opcode 0x%02x", op)
-	}
-	end, err := r.Byte()
-	if err != nil {
-		return constExprResult{}, fmt.Errorf("const expression missing end: %w", err)
-	}
-	if end != 0x0B {
-		return constExprResult{}, fmt.Errorf("const expression missing end")
-	}
-	if r.BytesLeft() != 0 {
-		return constExprResult{}, fmt.Errorf("const expression has trailing bytes")
-	}
-	if got.Type != want {
-		return constExprResult{}, fmt.Errorf("const expression type %s, want %s", got.Type, want)
-	}
-	return got, nil
-}
-
 // Signature returns the parameter and result types of an exported function.
 func (c *Compiled) Signature(export string) (params, results []wasm.ValType, err error) {
 	li, err := c.localIndex(export)
@@ -418,40 +156,6 @@ func (c *Compiled) localIndex(export string) (int, error) {
 		return 0, fmt.Errorf("export %q is an imported function", export)
 	}
 	return li, nil
-}
-
-func (c *Compiled) importedGlobalBits(imports Imports) ([]uint64, error) {
-	bits := make([]uint64, len(c.GlobalImports))
-	for i, imp := range c.GlobalImports {
-		key := imp.Module + "." + imp.Name
-		provided, ok := imports.Globals[key]
-		if !ok {
-			return nil, fmt.Errorf("missing imported global %q", key)
-		}
-		if provided.Type != imp.Type {
-			return nil, fmt.Errorf("imported global %q has type %s, want %s", key, provided.Type, imp.Type)
-		}
-		if provided.Mutable != imp.Mutable {
-			return nil, fmt.Errorf("imported global %q mutability mismatch", key)
-		}
-		bits[i] = normalizeGlobalBits(imp.Type, provided.Bits)
-	}
-	return bits, nil
-}
-
-func normalizeGlobalBits(t wasm.ValType, bits uint64) uint64 {
-	if t == wasm.I32 || t == wasm.F32 {
-		return uint64(uint32(bits))
-	}
-	return bits
-}
-
-func readGlobalSlot(globals []byte, idx int, t wasm.ValType) uint64 {
-	return normalizeGlobalBits(t, binary.LittleEndian.Uint64(globals[idx*8:]))
-}
-
-func writeGlobalSlot(globals []byte, idx int, t wasm.ValType, bits uint64) {
-	binary.LittleEndian.PutUint64(globals[idx*8:], normalizeGlobalBits(t, bits))
 }
 
 const wagoMagic = "WAGO"
@@ -493,191 +197,6 @@ func Load(b []byte) (*Compiled, error) {
 		return c, c.UnmarshalBinary(b)
 	}
 	return Compile(b)
-}
-
-// Instance is ready for repeated Invoke calls.
-type Instance struct {
-	c                      *Compiled
-	eng                    *runtime.Engine
-	jm                     *runtime.JobMemory
-	ar                     *runtime.Arena
-	base                   uintptr
-	mem                    []byte
-	hosts                  map[string]HostFunc
-	hostLog                []byte
-	globals                []byte
-	serArgs, results, trap []byte
-}
-
-// Instantiate maps code, initializes memory/table state, and allocates call buffers.
-func Instantiate(c *Compiled, hosts map[string]HostFunc) (*Instance, error) {
-	return InstantiateWithImports(c, Imports{Funcs: hosts})
-}
-
-// InstantiateWithImports maps code and supplies host functions and globals.
-func InstantiateWithImports(c *Compiled, imports Imports) (*Instance, error) {
-	importGlobalBits, err := c.importedGlobalBits(imports)
-	if err != nil {
-		return nil, err
-	}
-	eng, err := runtime.NewEngine()
-	if err != nil {
-		return nil, err
-	}
-	jm, err := runtime.NewJobMemory(1 << 16)
-	if err != nil {
-		eng.Close()
-		return nil, err
-	}
-	ar, err := runtime.NewArena(1 << 20)
-	if err != nil {
-		jm.Close()
-		eng.Close()
-		return nil, err
-	}
-	mem, base, err := runtime.MapCode(c.Code)
-	if err != nil {
-		ar.Close()
-		jm.Close()
-		eng.Close()
-		return nil, err
-	}
-	success := false
-	defer func() {
-		if success {
-			return
-		}
-		runtime.Unmap(mem)
-		ar.Close()
-		jm.Close()
-		eng.Close()
-	}()
-	const maxEntries = (1 << 16) / 8
-	hostLog := ar.Alloc(8 + maxEntries*8)
-	jm.SetCustomCtx(uintptr(unsafe.Pointer(&hostLog[0])))
-
-	var globals []byte
-	if len(c.Globals) > 0 {
-		globals = ar.Alloc(8 * len(c.Globals))
-		for i, g := range c.Globals {
-			bits := g.Bits
-			if i < len(importGlobalBits) {
-				bits = importGlobalBits[i]
-			} else if g.HasInitGlobal {
-				if g.InitGlobal < 0 || g.InitGlobal >= i {
-					return nil, fmt.Errorf("global %d initializer references unavailable global %d", i, g.InitGlobal)
-				}
-				bits = readGlobalSlot(globals, g.InitGlobal, g.Type)
-			}
-			writeGlobalSlot(globals, i, g.Type, bits)
-		}
-		jm.SetGlobalsPtr(uintptr(unsafe.Pointer(&globals[0])))
-	}
-
-	// Table descriptor: [len u32][pad][entry...], entry {codePtr u64, sigID u32, pad u32}.
-	if c.TableSize > 0 || len(c.Elems) > 0 {
-		size := c.TableSize
-		desc := ar.Alloc(8 + size*16)
-		binary.LittleEndian.PutUint32(desc, uint32(size))
-		for seg, el := range c.Elems {
-			elemBase := el.Base
-			if el.HasOffsetGlobal {
-				if el.OffsetGlobal < 0 || el.OffsetGlobal >= len(c.Globals) || el.OffsetGlobal*8+8 > len(globals) {
-					return nil, fmt.Errorf("element offset global %d out of range", el.OffsetGlobal)
-				}
-				elemBase = uint32(readGlobalSlot(globals, el.OffsetGlobal, c.Globals[el.OffsetGlobal].Type))
-			}
-			end := uint64(elemBase) + uint64(len(el.Funcs))
-			if end > uint64(size) {
-				return nil, fmt.Errorf("active element segment %d out of bounds: offset %d + length %d > table size %d", seg, elemBase, len(el.Funcs), size)
-			}
-			for k, fidx := range el.Funcs {
-				slot := int(elemBase) + k
-				off := 8 + slot*16
-				if li := int(fidx) - c.NumImports; li >= 0 && li < len(c.Entry) {
-					binary.LittleEndian.PutUint64(desc[off:], uint64(base)+uint64(c.Entry[li]))
-				}
-				if int(fidx) < len(c.FuncTypeID) {
-					binary.LittleEndian.PutUint32(desc[off+8:], c.FuncTypeID[fidx])
-				}
-			}
-		}
-		jm.SetTablePtr(uintptr(unsafe.Pointer(&desc[0])))
-	}
-
-	if len(c.Data) > 0 {
-		lin := jm.LinearMemory()
-		for seg, d := range c.Data {
-			off := d.Offset
-			if d.HasOffsetGlobal {
-				if d.OffsetGlobal < 0 || d.OffsetGlobal >= len(c.Globals) || d.OffsetGlobal*8+8 > len(globals) {
-					return nil, fmt.Errorf("data offset global %d out of range", d.OffsetGlobal)
-				}
-				off = uint32(readGlobalSlot(globals, d.OffsetGlobal, c.Globals[d.OffsetGlobal].Type))
-			}
-			end := uint64(off) + uint64(len(d.Bytes))
-			if end > uint64(len(lin)) {
-				return nil, fmt.Errorf("active data segment %d out of bounds: offset %d + length %d > memory size %d", seg, off, len(d.Bytes), len(lin))
-			}
-			copy(lin[off:end], d.Bytes)
-		}
-	}
-
-	success = true
-	return &Instance{
-		c: c, eng: eng, jm: jm, ar: ar, base: base, mem: mem, hosts: imports.Funcs, hostLog: hostLog, globals: globals,
-		serArgs: ar.Alloc(512), results: ar.Alloc(512), trap: ar.Alloc(8),
-	}, nil
-}
-
-// Close releases the instance's mapped code and memory.
-func (in *Instance) Close() {
-	runtime.Unmap(in.mem)
-	in.ar.Close()
-	in.jm.Close()
-	in.eng.Close()
-}
-
-// LinearMemory exposes the instance's linear memory for zero-copy access.
-func (in *Instance) LinearMemory() []byte { return in.jm.LinearMemory() }
-
-// Global returns the current value of an exported global.
-func (in *Instance) Global(name string) (Value, error) {
-	idx, ok := in.c.GlobalExports[name]
-	if !ok {
-		if _, isFunc := in.c.Exports[name]; isFunc {
-			return Value{}, fmt.Errorf("export %q is a function, not a global", name)
-		}
-		return Value{}, fmt.Errorf("no exported global %q", name)
-	}
-	if idx < 0 || idx >= len(in.c.Globals) || idx*8+8 > len(in.globals) {
-		return Value{}, fmt.Errorf("exported global %q index %d out of range", name, idx)
-	}
-	g := in.c.Globals[idx]
-	return Value{Type: g.Type, Bits: readGlobalSlot(in.globals, idx, g.Type)}, nil
-}
-
-// SetGlobal updates an exported mutable global.
-func (in *Instance) SetGlobal(name string, v Value) error {
-	idx, ok := in.c.GlobalExports[name]
-	if !ok {
-		if _, isFunc := in.c.Exports[name]; isFunc {
-			return fmt.Errorf("export %q is a function, not a global", name)
-		}
-		return fmt.Errorf("no exported global %q", name)
-	}
-	if idx < 0 || idx >= len(in.c.Globals) || idx*8+8 > len(in.globals) {
-		return fmt.Errorf("exported global %q index %d out of range", name, idx)
-	}
-	g := in.c.Globals[idx]
-	if !g.Mutable {
-		return fmt.Errorf("exported global %q is immutable", name)
-	}
-	if v.Type != g.Type {
-		return fmt.Errorf("exported global %q has type %s, got %s", name, g.Type, v.Type)
-	}
-	writeGlobalSlot(in.globals, idx, g.Type, v.Bits)
-	return nil
 }
 
 // Invoke marshals slot-based arguments/results around one native WasmWrapper call.
